Extract shared logic to gmae-configure-slice
