+++ app/src/store/game-config-slice.ts
@@ -236,6 +236,22 @@ const gameConfigSlice = createSlice({
   },
 });
 
+export function selectRoleSlots(state: GameConfigState): RoleSlot[] {
+  const { roleConfigMode, roleCounts, roleMins, roleMaxes } = state;
+  if (roleConfigMode === RoleConfigMode.Advanced) {
+    return Object.keys(roleMins)
+      .filter((id) => (roleMaxes[id] ?? 0) > 0)
+      .map((id) => ({
+        roleId: id,
+        min: roleMins[id] ?? 0,
+        max: roleMaxes[id] ?? 0,
+      }));
+  }
+  return Object.entries(roleCounts)
+    .filter(([, count]) => count > 0)
+    .map(([roleId, count]) => ({ roleId, min: count, max: count }));
+}
+
 export const {
   loadConfig,
   setGameMode,
